scara_socket/requirements.txt
python-socketio == 5.5.2
uvicorn[standard] == 0.17.6
opencv-contrib-python == 4.5.5.64
pyserial == 3.5
"python-socketio[client]"